requirements.txt
streamlit
fastapi
uvicorn
openai>=1.0.0
python-dotenv
requests
sqlalchemy
transformers
torch